Modify home page and correct back button access
