--- lib/screens/screen_content.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ScreenContent extends StatelessWidget {
-  const ScreenContent({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Screen 2'),
-      ),
-      body: SafeArea(
-        child: Column(
-          children: [
-            Text('Text on Screen 2'),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: Text('Go Back'),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
