+++ lib/screens/screen_home.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:test_proj/screens/screen_login.dart';
+
+class ScreenHome extends StatelessWidget {
+  const ScreenHome({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Home'),
+        actions: [
+          IconButton(
+            onPressed: () {
+              signout(context);
+            },
+            icon: Icon(Icons.logout),
+          ),
+        ],
+      ),
+      body: SafeArea(
+        child: Column(
+          children: [
+            Center(
+              child: Text('Text on Home'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  signout(BuildContext ctx) {
+    Navigator.of(ctx).pushAndRemoveUntil(MaterialPageRoute(builder: (ctx1) {
+      return ScreenLogin();
+    }), (route) => false);
+  }
+}
